backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
deepgram-sdk>=3.2.0
groq>=0.11.0
python-dotenv>=1.0.0
twilio>=9.0.0
httpx>=0.27.0
pydantic>=2.7.0
